🚚 (api): move target-profile-history-repository to certification shared context
